ref: Remove EMS\CommonBundle\Common\Standard\Hash
