+++ EMS/common-bundle/src/Common/CoreApi/TokenStore.php
@@ -2,7 +2,7 @@
 
 namespace EMS\CommonBundle\Common\CoreApi;
 
-use EMS\CommonBundle\Common\Standard\Hash;
+use EMS\Helpers\Standard\Hash;
 use Psr\Cache\CacheItemInterface;
 use Psr\Cache\CacheItemPoolInterface;
