+++ EMS/common-bundle/src/Twig/SearchRuntime.php
@@ -8,9 +8,9 @@ use Elastica\Query\BoolQuery;
 use Elastica\Query\Exists;
 use Elastica\Query\Nested;
 use Elastica\Result;
-use EMS\CommonBundle\Common\Standard\Hash;
 use EMS\CommonBundle\Search\Search;
 use EMS\CommonBundle\Service\ElasticaService;
+use EMS\Helpers\Standard\Hash;
 use Twig\Extension\RuntimeExtensionInterface;
 
 final class SearchRuntime implements RuntimeExtensionInterface
